Restart this project.
I'm using the DCPU Toolchain (http://dcputoolcha.in/), with m4 for a
preprocessor. The toolchain comes with a preprocessor of its own, but
it just wasn't powerful enough to make assembly programming bearable
for me.

From the looks of things, I'll be contributing patches for the
assembler as I fix various issues. It'll be a while before this Forth
is ready to go.